feat: implement BlockchainService and SigningService for blockchain operations and cryptographic signing
