+++ Blockchain-Demo-Project/Blockchain.cs
@@ -35,7 +35,7 @@ public class Blockchain : IBlockchain
         PendingTransactions.Clear();
     }
 
-    public void AddBlock(IBlock block)
+    public virtual void AddBlock(IBlock block)
     {
         if (block == null) throw new ArgumentNullException(nameof(block));
         if (!block.ValidBlock(Difficulty)) throw new InvalidOperationException("Invalid block.");
@@ -78,7 +78,7 @@ public class Blockchain : IBlockchain
         }
     }
 
-    public void AddTransaction(ITransact transaction)
+    public virtual void AddTransaction(ITransact transaction)
     {
         var verify = VerifyTransaction(transaction);
         if (!verify)
@@ -120,7 +120,7 @@ public class Blockchain : IBlockchain
     }
 }
 
-public class TestChain : IBlockchain
+public class TestChain : Blockchain, IBlockchain
 {
     public string Name { get; }
     public int Difficulty { get; }
@@ -144,56 +144,8 @@ public class TestChain : IBlockchain
     private List<IBlock> Chain { get;  } = new();
     public IReadOnlyList<IBlock> GetChain() => Chain.AsReadOnly();
     public IBlock GetLatestBlock() => GetChain().Last();
-    public void AddBlock(IBlock block) { }
-    public void AddTransaction(ITransact transaction) { }
+    public override void AddBlock(IBlock block) { }
+    public override void AddTransaction(ITransact transaction) { }
     public decimal GetBalance(string walletAddress) => 0;
     public void ValidateChain() { }
-}
-
-public class BlockchainService: IBlockchainService
-{
-    private IBlockchain Blockchain { get;  }
-
-    private BlockchainService(IBlockchain blockchain)
-    {
-        Blockchain = blockchain ?? throw new ArgumentNullException(nameof(blockchain));
-    }
-
-    public static BlockchainService Create(IBlockchain blockchain)
-    {
-        return new BlockchainService(blockchain);
-    }
-
-    private void ExecuteMiner(string address)
-    {
-        var miner = Miner.Create(address);
-        var minerThread = new Thread(() => miner.MineBlock(Blockchain));
-        minerThread.Start();
-    }
-
-    public void InitializeBlockchain(string address)
-    {
-        if (GetChain().Count == 1)
-        {
-            ExecuteMiner(address);
-        }
-    }
-
-    public void AddTransaction(ITransact transaction, string privateKey)
-    {
-        if (transaction == null) throw new ArgumentNullException(nameof(transaction));
-        if (string.IsNullOrEmpty(privateKey)) throw new ArgumentException("Private key cannot be null or empty.", nameof(privateKey));
-        transaction.SignTransaction(privateKey);
-        Blockchain.AddTransaction(transaction);
-        ExecuteMiner(transaction.FromAddress);
-    }
-
-    public decimal GetBalance(string walletAddress)
-    {
-        return Blockchain.GetBalance(walletAddress);
-    }
-
-    public IReadOnlyList<IBlock> GetChain() => Blockchain.GetChain();
-
-    public string GetChainName() => Blockchain.Name;
 }
